[REFACTOR]: createStore is more robust for power users
